ai-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0

# AI/ML dependencies
openai==1.3.7
sentence-transformers==2.2.2
pinecone-client==2.2.4
numpy==1.24.4
pandas==2.1.4

# Optional: for better performance
aiofiles==23.2.1
redis==5.0.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1